Modified config file and status example
The link to the test result is simulator dependant. Thus the link
to the result has to be moved to the simulator column.

Expanded and renamed the configuration file. We need a uniq name
for each simulator. The simulator type is required to select the
correct test files and the folder is required as prefix for the
resulting test file.